backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.32
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
Werkzeug==2.0.1
PyJWT==2.4.0
requests==2.28.2